fixed 2 tests (hopefully); fixed linux warnings
git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/branches/Release1.4@5020 6adb6e08-d915-0410-941f-83917bcadc18

+++ include/OpenMS/TRANSFORMATIONS/RAW2PEAK/ContinuousWaveletTransformNumIntegration.h
@@ -146,7 +146,7 @@ namespace OpenMS
 				else processed_input[0]=it_help->getIntensity();
 				
         double x;
-        for (SignedSize k=1; k < n-zeros; ++k)
+        for (SignedSize k=1; k < n-(int)zeros; ++k)
         {
           x = origin + k*spacing;
           // go to the real data point next to x
@@ -162,7 +162,7 @@ namespace OpenMS
 				}
 
         // TODO avoid to compute the cwt for the zeros in signal
-        for (UInt i=0; i < n; ++i)
+        for (Int i=0; i < n; ++i)
         {
           signal_[i].setMZ(origin + i*spacing);
 					signal_[i].setIntensity((Peak1D::IntensityType)integrate_(processed_input,spacing,i));
